Code cleanup, fix warnings and messages

+++ src/Core/FSpot/Utils/TargetListExtensionMethods.cs
@@ -13,12 +13,12 @@ using Gtk;
 
 namespace FSpot.Utils
 {
-    public static class TargetListExtensionMethods
-    {
-        public static void AddTargetEntry (this TargetList targetList, TargetEntry entry)
-        {
-            targetList.Add (entry.Target, (uint)entry.Flags, (uint)entry.Info);
-        }
-    }
+	public static class TargetListExtensionMethods
+	{
+		public static void AddTargetEntry (this TargetList targetList, TargetEntry entry)
+		{
+			targetList.Add (entry.Target, (uint)entry.Flags, entry.Info);
+		}
+	}
 }
